#348 Various improvements and refactorings.
