+++ platform-pojo-bl/src/test/java/ua/com/fielden/platform/types/RichTextMarkdownSanitizationTest.java
@@ -142,18 +142,18 @@ public class RichTextMarkdownSanitizationTest {
     }
 
     private void assertSanitizationFailure(final String input) {
-        assertFalse(RichText.sanitizeMarkdown(input).isSuccessful());
+        assertFalse(RichText.sanitiseMarkdown(input).isSuccessful());
     }
 
     private void assertSanitizationSuccess(final String input) {
-        final Result result = RichText.sanitizeMarkdown(input);
+        final Result result = RichText.sanitiseMarkdown(input);
         if (!result.isSuccessful()) {
             fail(result.getMessage());
         }
     }
 
     private void assertAfterSanitization(final String expected, final String input) {
-        final Result result = RichText.sanitizeMarkdown(input);
+        final Result result = RichText.sanitiseMarkdown(input);
         if (!result.isSuccessful()) {
             fail(result.getMessage());
         } else {
